Version 1.0.3 Moved to JAXB 3.0.0 Jakarta update. Removed executable JAR features. Moved main classes to src/test.
