+++ src/main/java/nl/rug/ds/bpm/specification/jaxb/Condition.java
@@ -1,7 +1,7 @@
 package nl.rug.ds.bpm.specification.jaxb;
 
-import javax.xml.bind.annotation.XmlRootElement;
-import javax.xml.bind.annotation.XmlValue;
+import jakarta.xml.bind.annotation.XmlRootElement;
+import jakarta.xml.bind.annotation.XmlValue;
 
 /**
  * Created by p256867 on 6-4-2017.
