+++ src/main/java/nl/rug/ds/bpm/specification/jaxb/Element.java
@@ -1,7 +1,7 @@
 package nl.rug.ds.bpm.specification.jaxb;
 
-import javax.xml.bind.annotation.XmlRootElement;
-import javax.xml.bind.annotation.XmlValue;
+import jakarta.xml.bind.annotation.XmlRootElement;
+import jakarta.xml.bind.annotation.XmlValue;
 
 /**
  * Created by Heerko Groefsema on 07-Apr-17.
