+++ src/main/java/nl/rug/ds/bpm/specification/marshaller/SpecificationMarshaller.java
@@ -3,9 +3,9 @@ package nl.rug.ds.bpm.specification.marshaller;
 import nl.rug.ds.bpm.specification.jaxb.BPMSpecification;
 import nl.rug.ds.bpm.util.exception.SpecificationException;
 
-import javax.xml.bind.JAXBContext;
-import javax.xml.bind.JAXBException;
-import javax.xml.bind.Marshaller;
+import jakarta.xml.bind.JAXBContext;
+import jakarta.xml.bind.JAXBException;
+import jakarta.xml.bind.Marshaller;
 import java.io.File;
 import java.io.OutputStream;
