+++ src/main/java/nl/rug/ds/bpm/specification/marshaller/SpecificationUnmarshaller.java
@@ -3,8 +3,8 @@ package nl.rug.ds.bpm.specification.marshaller;
 import nl.rug.ds.bpm.specification.jaxb.BPMSpecification;
 import nl.rug.ds.bpm.util.exception.SpecificationException;
 
-import javax.xml.bind.JAXBContext;
-import javax.xml.bind.Unmarshaller;
+import jakarta.xml.bind.JAXBContext;
+import jakarta.xml.bind.Unmarshaller;
 import java.io.File;
 import java.io.InputStream;
