+++ src/test/java/nl/rug/ds/bpm/test/InteractiveVerifier.java
@@ -1,4 +1,4 @@
-package nl.rug.ds.bpm.main;
+package nl.rug.ds.bpm.test;
 
 import nl.rug.ds.bpm.petrinet.interfaces.net.VerifiableNet;
 import nl.rug.ds.bpm.petrinet.ptnet.PlaceTransitionNet;
